+++ src/main/kotlin/fi/elsapalvelu/elsa/scheduler/jobs/ScheduledSisuTutkintoohjelmaImport.kt
@@ -1,8 +1,8 @@
 package fi.elsapalvelu.elsa.scheduler.jobs
 
 import fi.elsapalvelu.elsa.scheduler.AbstractTriggerableJob
-import fi.elsapalvelu.elsa.service.SisuTutkintoohjelmaFetchingService
-import fi.elsapalvelu.elsa.service.SisuTutkintoohjelmaImportService
+import fi.elsapalvelu.elsa.service.integration.sisu.SisuTutkintoohjelmaFetchingService
+import fi.elsapalvelu.elsa.service.integration.sisu.SisuTutkintoohjelmaImportService
 import kotlinx.coroutines.runBlocking
 import net.javacrumbs.shedlock.spring.annotation.SchedulerLock
 import org.springframework.scheduling.annotation.Scheduled
