--- src/main/kotlin/fi/elsapalvelu/elsa/service/SisuTutkintoohjelmaFetchingService.kt
@@ -1,7 +1,0 @@
-package fi.elsapalvelu.elsa.service
-
-import fi.elsapalvelu.elsa.service.impl.Qualifications
-
-interface SisuTutkintoohjelmaFetchingService {
-    suspend fun fetch(): Qualifications?
-}
