--- src/main/kotlin/fi/elsapalvelu/elsa/service/SisuTutkintoohjelmaImportService.kt
@@ -1,7 +1,0 @@
-package fi.elsapalvelu.elsa.service
-
-import fi.elsapalvelu.elsa.service.impl.Qualifications
-
-interface SisuTutkintoohjelmaImportService {
-    fun import(qualifications: Qualifications)
-}
