+++ src/main/kotlin/fi/elsapalvelu/elsa/service/integration/peppi/turku/PeppiTurkuClientBuilderImpl.kt
@@ -1,44 +1,42 @@
-package fi.elsapalvelu.elsa.service.impl
+package fi.elsapalvelu.elsa.service.integration.peppi.turku
 
 import fi.elsapalvelu.elsa.config.ApplicationProperties
 import fi.elsapalvelu.elsa.interceptor.OkHttp3RequestInterceptor
-import fi.elsapalvelu.elsa.security.AccessTokenAuthenticator
-import fi.elsapalvelu.elsa.service.AuthenticationTokenService
 import fi.elsapalvelu.elsa.service.OkHttpClientBuilder
 import okhttp3.OkHttpClient
 import org.springframework.beans.factory.annotation.Qualifier
 import org.springframework.stereotype.Service
 import java.util.concurrent.TimeUnit
 
-@Qualifier("SisuTre")
+@Qualifier("PeppiTurku")
 @Service
-class SisuTreClientBuilderImpl(
-    sisuTreAuthenticationTokenService: AuthenticationTokenService,
+class PeppiTurkuClientBuilderImpl(
     applicationProperties: ApplicationProperties
 ) : OkHttpClientBuilder {
 
     init {
         Companion.applicationProperties = applicationProperties
-        Companion.sisuTreAuthenticationTokenService = sisuTreAuthenticationTokenService
     }
 
     companion object {
 
         private lateinit var applicationProperties: ApplicationProperties
-        private lateinit var sisuTreAuthenticationTokenService: AuthenticationTokenService
 
         val okHttpClient: OkHttpClient by lazy {
             OkHttpClient.Builder()
                 .addInterceptor(
                     OkHttp3RequestInterceptor(
                         mapOf(
-                            "Ocp-Apim-Subscription-Key" to applicationProperties.getSecurity()
-                                .getSisuTre().subscriptionKey!!,
-                            "Content-Type" to "application/json"
+                            "Accept" to "application/json",
+                            "ESP-ScreenName" to "peppi_elsa",
+                            "username" to "peppi_elsa",
+                            "X-Api-Key" to applicationProperties.getSecurity().getPeppiTurku().apiKey!!,
+                            "Authorization" to "Basic ${
+                                applicationProperties.getSecurity().getPeppiTurku().basicAuthEncodedKey!!
+                            }"
                         )
                     )
                 )
-                .authenticator(AccessTokenAuthenticator(sisuTreAuthenticationTokenService))
                 .connectTimeout(5, TimeUnit.SECONDS)
                 .readTimeout(5, TimeUnit.SECONDS)
                 .writeTimeout(5, TimeUnit.SECONDS)
@@ -47,10 +45,6 @@ class SisuTreClientBuilderImpl(
     }
 
     override fun okHttpClient(): OkHttpClient {
-        return okHttpClient.newBuilder().addInterceptor(
-            OkHttp3RequestInterceptor(
-                mapOf("Authorization" to "Bearer ${sisuTreAuthenticationTokenService.getCachedTokenOrRequestNew()}")
-            )
-        ).build()
+        return okHttpClient
     }
 }
