+++ src/main/kotlin/fi/elsapalvelu/elsa/service/integration/peppi/turku/PeppiTurkuOpintosuorituksetFetchingServiceImpl.kt
@@ -1,25 +1,27 @@
-package fi.elsapalvelu.elsa.service.impl
+package fi.elsapalvelu.elsa.service.integration.peppi.turku
 
 import fi.elsapalvelu.elsa.config.ApplicationProperties
 import fi.elsapalvelu.elsa.domain.enumeration.YliopistoEnum
 import fi.elsapalvelu.elsa.repository.YliopistoRepository
 import fi.elsapalvelu.elsa.service.OkHttpClientBuilder
-import fi.elsapalvelu.elsa.service.OpintotietodataFetchingService
-import fi.elsapalvelu.elsa.service.dto.OpintotietodataDTO
+import fi.elsapalvelu.elsa.service.OpintosuorituksetFetchingService
+import fi.elsapalvelu.elsa.service.integration.peppi.PeppiCommonOpintosuorituksetFetchingService
+import fi.elsapalvelu.elsa.service.dto.OpintosuorituksetPersistenceDTO
 import org.springframework.beans.factory.annotation.Qualifier
 import org.springframework.stereotype.Service
 
 @Service
-class PeppiTurkuOpintotietodataFetchingServiceImpl(
+class PeppiTurkuOpintosuorituksetFetchingServiceImpl(
     @Qualifier("PeppiTurku") private val peppiTurkuClientBuilder: OkHttpClientBuilder,
-    private val commonOpintotietodataFetchingServiceImpl: PeppiCommonOpintotietodataFetchingServiceImpl,
+    private val commonOpintosuorituksetFetchingService: PeppiCommonOpintosuorituksetFetchingService,
     private val applicationProperties: ApplicationProperties,
     private val yliopistoRepository: YliopistoRepository
-) : OpintotietodataFetchingService {
+) : OpintosuorituksetFetchingService {
 
-    override suspend fun fetchOpintotietodata(hetu: String): OpintotietodataDTO? {
-        val endpointBaseUrl = "${applicationProperties.getSecurity().getPeppiTurku().endpointUrl!!}/student"
-        return commonOpintotietodataFetchingServiceImpl.fetchOpintotietodata(
+    override suspend fun fetchOpintosuoritukset(hetu: String): OpintosuorituksetPersistenceDTO? {
+        val endpointBaseUrl =
+            "${applicationProperties.getSecurity().getPeppiTurku().endpointUrl!!}/study_accomplishments"
+        return commonOpintosuorituksetFetchingService.fetchOpintosuoritukset(
             endpointBaseUrl,
             peppiTurkuClientBuilder.okHttpClient(),
             hetu,
@@ -27,7 +29,7 @@ class PeppiTurkuOpintotietodataFetchingServiceImpl(
         )
     }
 
-    override fun shouldFetchOpintotietodata(): Boolean {
+    override fun shouldFetchOpintosuoritukset(): Boolean {
         return yliopistoRepository.findOneByNimi(YliopistoEnum.TURUN_YLIOPISTO)?.haeOpintotietodata == true
     }
