+++ src/main/kotlin/fi/elsapalvelu/elsa/service/integration/sisu/SisuTutkintoohjelmaFetchingService.kt
@@ -1,0 +1,7 @@
+package fi.elsapalvelu.elsa.service.integration.sisu
+
+import fi.elsapalvelu.elsa.service.integration.sisu.Qualifications
+
+interface SisuTutkintoohjelmaFetchingService {
+    suspend fun fetch(): Qualifications?
+}
