+++ src/main/kotlin/fi/elsapalvelu/elsa/service/integration/sisu/SisuTutkintoohjelmaImportServiceImpl.kt
@@ -1,8 +1,8 @@
-package fi.elsapalvelu.elsa.service.impl
+package fi.elsapalvelu.elsa.service.integration.sisu
 
 import fi.elsapalvelu.elsa.domain.ErikoisalaSisuTutkintoohjelma
 import fi.elsapalvelu.elsa.repository.ErikoisalaRepository
-import fi.elsapalvelu.elsa.service.SisuTutkintoohjelmaImportService
+import fi.elsapalvelu.elsa.service.integration.sisu.SisuTutkintoohjelmaImportService
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
