+++ src/main/kotlin/fi/elsapalvelu/elsa/service/integration/sisu/helsinki/SisuHyOpintosuorituksetFetchingServiceImpl.kt
@@ -1,6 +1,6 @@
-package fi.elsapalvelu.elsa.service.impl
+package fi.elsapalvelu.elsa.service.integration.sisu.helsinki
 
-import fi.elsapalvelu.elsa.OpintosuorituksetPeppiOuluQuery
+import fi.elsapalvelu.elsa.OpintosuorituksetSisuHyQuery
 import fi.elsapalvelu.elsa.domain.enumeration.YliopistoEnum
 import fi.elsapalvelu.elsa.extensions.checkErrors
 import fi.elsapalvelu.elsa.extensions.tryParseToLocalDate
@@ -14,61 +14,66 @@ import org.slf4j.LoggerFactory
 import org.springframework.beans.factory.annotation.Qualifier
 import org.springframework.stereotype.Service
 
+private val validStates = listOf("ATTAINED", "SUBSTITUTED")
+
 @Service
-class PeppiOuluOpintosuorituksetFetchingServiceImpl(
-    @Qualifier("PeppiOulu") private val peppiOuluClientBuilder: GraphQLClientBuilder,
+class SisuHyOpintosuorituksetFetchingServiceImpl(
+    @Qualifier("SisuHy") private val sisuHyClientBuilder: GraphQLClientBuilder,
     private val yliopistoRepository: YliopistoRepository
 ) : OpintosuorituksetFetchingService {
 
     private val log = LoggerFactory.getLogger(javaClass)
 
     override suspend fun fetchOpintosuoritukset(hetu: String): OpintosuorituksetPersistenceDTO? {
-        val response = peppiOuluClientBuilder.apolloClient()
-            .query(OpintosuorituksetPeppiOuluQuery(id = hetu))
+        val response = sisuHyClientBuilder.apolloClient()
+            .query(OpintosuorituksetSisuHyQuery(id = hetu))
             .execute()
-            .checkErrors("Opintosuoritustietoja ei saatu haettua Oulun Pepistä", log)
+            .checkErrors("Opintosuoritustietoja ei saatu haettua HY:n Sisusta", log)
 
-        return response.data?.private_person_by_personal_identity_code?.attainments?.let {
+        return response.data?.private_person_by_personal_identity_code?.attainments?.filter {
+            it.state in validStates
+        }?.let {
             OpintosuorituksetPersistenceDTO(
-                yliopisto = YliopistoEnum.OULUN_YLIOPISTO,
+                yliopisto = YliopistoEnum.HELSINGIN_YLIOPISTO,
                 items = it.map { a ->
                     OpintosuoritusDTO(
                         suorituspaiva = a.attainmentDate?.tryParseToLocalDate(),
                         opintopisteet = a.credits,
-                        nimi_fi = a.courseUnit?.name?.fi,
-                        nimi_sv = a.courseUnit?.name?.sv,
-                        kurssikoodi = a.courseUnit?.code,
+                        nimi_fi = a.name?.fi ?: a.courseUnit?.name?.fi ?: a.module?.name?.fi,
+                        nimi_sv = a.name?.sv ?: a.courseUnit?.name?.sv ?: a.module?.name?.fi,
+                        kurssikoodi = a.code ?: a.courseUnit?.code ?: a.module?.code,
                         hyvaksytty = a.grade?.passed,
                         arvio_fi = a.grade?.name?.fi,
                         arvio_sv = a.grade?.name?.sv,
                         vanhenemispaiva = a.expiryDate?.tryParseToLocalDate(),
                         yliopistoOpintooikeusId = a.studyRightId,
-                        osakokonaisuudet = a.childAttainments?.map { c -> mapOsakokonaisuus(c) }
+                        osakokonaisuudet = a.childAttainments?.filter { c -> c.state in validStates }
+                            ?.map { c -> mapOsakokonaisuus(c) }
                     )
                 }
             )
         }
     }
 
-    private fun mapOsakokonaisuus(osakokonaisuus: OpintosuorituksetPeppiOuluQuery.ChildAttainment): OpintosuoritusOsakokonaisuusDTO {
+    private fun mapOsakokonaisuus(osakokonaisuus: OpintosuorituksetSisuHyQuery.ChildAttainment): OpintosuoritusOsakokonaisuusDTO {
         return OpintosuoritusOsakokonaisuusDTO(
             suorituspaiva = osakokonaisuus.attainmentDate?.tryParseToLocalDate(),
             opintopisteet = osakokonaisuus.credits,
-            nimi_fi = osakokonaisuus.courseUnit?.name?.fi,
-            nimi_sv = osakokonaisuus.courseUnit?.name?.sv,
-            kurssikoodi = osakokonaisuus.courseUnit?.code,
+            nimi_fi = osakokonaisuus.name?.fi ?: osakokonaisuus.courseUnit?.name?.fi,
+            nimi_sv = osakokonaisuus.name?.sv ?: osakokonaisuus.courseUnit?.name?.sv,
+            kurssikoodi = osakokonaisuus.code ?: osakokonaisuus.courseUnit?.code,
             hyvaksytty = osakokonaisuus.grade?.passed,
             arvio_fi = osakokonaisuus.grade?.name?.fi,
             arvio_sv = osakokonaisuus.grade?.name?.sv,
-            vanhenemispaiva = osakokonaisuus.expiryDate?.tryParseToLocalDate(),
+            vanhenemispaiva = osakokonaisuus.expiryDate?.tryParseToLocalDate()
         )
     }
 
     override fun shouldFetchOpintosuoritukset(): Boolean {
-        return yliopistoRepository.findOneByNimi(YliopistoEnum.OULUN_YLIOPISTO)?.haeOpintotietodata == true
+        return yliopistoRepository.findOneByNimi(YliopistoEnum.HELSINGIN_YLIOPISTO)?.haeOpintotietodata == true
     }
 
     override fun getYliopisto(): YliopistoEnum {
-        return YliopistoEnum.OULUN_YLIOPISTO
+        return YliopistoEnum.HELSINGIN_YLIOPISTO
     }
 }
