+++ src/main/kotlin/fi/elsapalvelu/elsa/service/integration/sisu/helsinki/SisuHyOpintotietodataFetchingServiceImpl.kt
@@ -1,4 +1,4 @@
-package fi.elsapalvelu.elsa.service.impl
+package fi.elsapalvelu.elsa.service.integration.sisu.helsinki
 
 import fi.elsapalvelu.elsa.OpintotietodataSisuHyQuery
 import fi.elsapalvelu.elsa.config.ERIKOISTUVA_HAMMASLAAKARI_SISU_KOULUTUS
